Senior Bonuses

+++ Assets/Scripts/UI/MoneyDisplay.cs
@@ -5,7 +5,7 @@ public class MoneyDisplay : MonoBehaviour
 {
     [SerializeField] private Text moneyText;
     
-    public void SetMoney(float money)
+    public void SetMoney(double money)
     {
         moneyText.text = money.ToString("C");
     }
